Adds initial build of rough working version
Note: This initial build is entirely the result of ChatGPT
interactions. Yes it's a mess (in terms of visuals, code structure,
interactions, etc.). But -- it's entirely the result of ChatGPT
and... it actually works.

The entire process with ChatGPT to get this far (with many, many
back-and-forth iterations) took about 1.5 hrs.

Next steps are: clean it up.