WIP restructures code into classes
This is very much WIP. Currently this code will build most of the
existing UI using a UI object, and to manage the data with a combo
of an object per track, and a tracks manager.

State management for each beat (each circle) exists, but playback
and UI functionality for bars/beats/speed/play/pause/reset does not.