Refactors configuration into state object
We now have a single State class instance which is the source of
truth for all configuration and state of the sequencer.

Classes in the application reference the state instance directly
to get or set values, in their native variable type.

The State class can also export the state to a URL, and import
the state from a URL.

UI controls are dynamically created now and their defaults are
set to whatever the initial state is.